Updated during class 2/15/18

+++ Lab1/src/pkgCore/Die.java
@@ -2,11 +2,11 @@ package pkgCore;
 
 public class Die {
 
-	private int DieValue;
+	private byte DieValue;
 
 	public Die() {
 		// TODO: Determine DieVaue.. a random number between 1 and 6
-		DieValue = (int)(Math.random()*6 + 1);			
+		DieValue = (byte)(Math.random()*6 + 1);			
 	}
 
 	public int getDieValue() {
